http: use the request context
CloseNotifier has been deprecated for a while.

Also, ditch the "environment" context. We don't actually _need_ this.
